Add BigDecimal version
This adds a big decimal version of the algorithm.  At this stage,
however, it isn't getting much more precision.  Potentially it could be
improved with further tweaking to the various precision settings for
BigDecimal (e.g. via MathContext).  Also I have hard coded the sqrt(2)
with a fixed number of decimal places which could be improved also.